// backend/src/common/enums/auditoria.enum.ts
export enum AuditoriaAcao {
  CRIAR = 'CRIAR',
  ATUALIZAR = 'ATUALIZAR',
  ATIVAR = 'ATIVAR',
  DESATIVAR = 'DESATIVAR',
  EXCLUIR = 'EXCLUIR',
  LOGIN = 'LOGIN',
  ENTRADA_ESTOQUE = 'ENTRADA_ESTOQUE',
  SAIDA_ESTOQUE = 'SAIDA_ESTOQUE',
  AJUSTE_ESTOQUE = 'AJUSTE_ESTOQUE',
  INVENTARIO_ESTOQUE = 'INVENTARIO_ESTOQUE',
}

export enum AuditoriaEntidade {
  EMPRESA = 'EMPRESA',
  USUARIO = 'USUARIO',
  MODULO = 'MODULO',
  EMPRESA_MODULO = 'EMPRESA_MODULO',
  CATEGORIA_PRODUTO = 'CATEGORIA_PRODUTO',
  PRODUTO = 'PRODUTO',
  ESTOQUE = 'ESTOQUE',
  MOVIMENTACAO_ESTOQUE = 'MOVIMENTACAO_ESTOQUE',
}